requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
openai-whisper==20230918
python_multipart==0.0.6
psycopg2==2.9.7
python-dotenv==1.0.0